Do real workaround for fullscreencover. Enable iOS 18.5/macOS 15.5 again.
