--- Adventurers/Views/Browser/BrowserViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  BrowserViewModel.swift
-//  Adventurers
-//
-//  Created by Syd Polk on 6/11/25.
-//
-
-import Foundation
